+++ cordapp-cpk2/src/main/java/net/corda/plugins/cpb2/CpbPlugin.java
@@ -17,6 +17,7 @@ import org.jetbrains.annotations.NotNull;
 
 import static net.corda.plugins.cpk2.CordappUtils.ALL_CORDAPPS_CONFIGURATION_NAME;
 import static net.corda.plugins.cpk2.CordappUtils.copyJarEnabledTo;
+import static net.corda.plugins.cpk2.CordappUtils.setCannotBeDeclared;
 import static net.corda.plugins.cpk2.SignJar.sign;
 import static net.corda.plugins.cpk2.SigningProperties.nested;
 import static org.gradle.api.artifacts.Dependency.ARCHIVES_CONFIGURATION;
@@ -46,10 +47,12 @@ public final class CpbPlugin implements Plugin<Project> {
             .setVisible(false)
             .extendsFrom(cpbConfiguration);
         cpbPackaging.setCanBeConsumed(false);
+        setCannotBeDeclared(cpbPackaging);
 
-        project.getConfigurations().create(CORDA_CPB_CONFIGURATION_NAME)
-            .attributes(attributor::forCpb)
-            .setCanBeResolved(false);
+        final Configuration cordaCPB = project.getConfigurations().create(CORDA_CPB_CONFIGURATION_NAME)
+            .attributes(attributor::forCpb);
+        cordaCPB.setCanBeResolved(false);
+        setCannotBeDeclared(cordaCPB);
 
         final TaskProvider<Jar> cpkTask = project.getTasks().named(JAR_TASK_NAME, Jar.class);
         final Provider<RegularFile> cpkPath = cpkTask.flatMap(Jar::getArchiveFile);